fix: make dendrite pushes work
Dendrite is a tiny bit off spec here, by sending ms timestamps instead
of seconds. They don't fit in a 32 bit integer, which means dendrite
push notifications don't get through. To work around that and to be a
bit more resilient (as well as to make sure this continues working after
2038), this commit changes these timestamps to be stored with 64 bits
instead.

+++ src/models.rs
@@ -66,7 +66,7 @@ pub struct Device {
 	/// Push key
 	pub pushkey: String,
 	/// Timestamp of the last Push key update
-	pub pushkey_ts: Option<u32>,
+	pub pushkey_ts: Option<u64>,
 	/// Pusher specific data
 	pub data: Option<PusherData>,
 	/// A dictionary of customisations made to the way this notification is to
